requirements.txt
streamlit
streamlit_option_menu
gspread
google-auth
google-auth-oauthlib
pandas
oauth2client
streamlit_oauth
qrcode
pillow
google-api-python-client
requests
authlib>=1.2.1
numpy
plotly
google-auth-httplib2
openpyxl
qrcode[pil]
reportlab>=3.6.0
reportlab
kaleido